# api/docs/swagger.yaml
definitions:
  template_variables.CreateResponse:
    properties:
      id:
        type: string
    type: object
...
  template_variables.FailureResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  template_variables.SuccessResponse:
    properties:
      success:
        example: true
        type: boolean
    type: object
...
  user_service.GetAllPermissionsResponse:
    properties:
      count:
        type: integer
      permissions:
        items:
          $ref: '#/definitions/user_service.Permission'
        type: array
    type: object
  user_service.GetAllRolesResponse:
    properties:
      count:
        type: integer
      roles:
        items:
          $ref: '#/definitions/user_service.Role'
        type: array
    type: object
  user_service.Organization:
    properties:
      code:
        type: integer
      created_at:
        type: string
      description:
        type: string
      external_id:
        type: integer
      full_name:
        type: string
      id:
        type: string
      inn:
        type: string
      name:
        type: string
      responsible_orgs:
        items:
          $ref: '#/definitions/user_service.ResponsibleUser'
        type: array
      soato:
        type: integer
      status:
        type: boolean
      type:
        type: string
      updated_at:
        type: string
    type: object
  user_service.Permission:
    properties:
      description:
        type: string
      id:
        type: string
      label:
        type: string
      name:
        type: string
      ru_name:
        type: string
    type: object
  user_service.ResponsibleUser:
    properties:
      organization_id:
        type: string
      soato:
        type: string
    type: object
  user_service.Role:
    properties:
      code:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      organization:
        $ref: '#/definitions/user_service.Organization'
        type: object
      permissions:
        items:
          $ref: '#/definitions/user_service.Permission'
        type: array
      status:
        type: boolean
      updated_at:
        type: string
    type: object
...
  var_user_service.CreateUpdatePermissionSwag:
    properties:
      description:
        type: string
      label:
        type: string
      name:
        type: string
      ru_name:
        type: string
    required:
    - name
    - ru_name
    type: object
  var_user_service.CreateUpdateRoleSwag:
    properties:
      code:
        type: integer
      description:
        type: string
      name:
        type: string
      organization_id:
        type: string
      permissions:
        items:
          type: string
        type: array
      status:
        type: boolean
    required:
    - name
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /v1/permission:
    get:
      consumes:
      - application/json
      description: API for getting all rolees
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.GetAllPermissionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: Getting All permissions
      tags:
      - permission
    post:
      consumes:
      - application/json
      description: API for creating permission
      parameters:
      - description: permission
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/var_user_service.CreateUpdatePermissionSwag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/template_variables.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: Create permission
      tags:
      - permission
  /v1/permission/{permission_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: permission_id
        in: path
        name: permission_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Permission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: Get Permission
      tags:
      - permission
    put:
      consumes:
      - application/json
      description: API for updating permission
      parameters:
      - description: permission_id
        in: path
        name: permission_id
        required: true
        type: string
      - description: permission
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/var_user_service.CreateUpdatePermissionSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/template_variables.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: Update permission
      tags:
      - permission
  /v1/role:
    get:
      consumes:
      - application/json
      description: API for getting all rolees
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: organization_id
        in: query
        name: organization_id
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.GetAllRolesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: Getting All roles
      tags:
      - role
    post:
      consumes:
      - application/json
      description: API for creating role
      parameters:
      - description: role
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/var_user_service.CreateUpdateRoleSwag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/template_variables.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: Create role
      tags:
      - role
  /v1/role/{role_id}:
    delete:
      consumes:
      - application/json
      description: API for deleting role
      parameters:
      - description: role_id
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/template_variables.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: Delete Role
      tags:
      - role
    get:
      consumes:
      - application/json
      description: API for getting role
      parameters:
      - description: role_id
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: Get role
      tags:
      - role
    put:
      consumes:
      - application/json
      description: API for updating role
      parameters:
      - description: role_id
        in: path
        name: role_id
        required: true
        type: string
      - description: role
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/var_user_service.CreateUpdateRoleSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/template_variables.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: Update role
      tags:
      - role
  /v1/role/code:
    get:
      consumes:
      - application/json
      description: API for getting all rolees by code
      parameters:
      - description: code
        in: query
        name: code
        type: integer
      - description: organization_id
        in: query
        name: organization_id
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.GetAllRolesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/template_variables.FailureResponse'
      summary: Getting All roles by code
      tags:
      - role
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"